refactor: RestController 적용
