+++ src/test/resources/sql/ddl.sql
@@ -1,5 +1,4 @@
-set
-mode MySQL;
+set mode MySQL;
 
 DROP TABLE rooms IF EXISTS;
